Move MCP server config tests to codex-config
