+++ codex-rs/config/src/mcp_types.rs
@@ -325,3 +325,7 @@ mod option_duration_secs {
             .transpose()
     }
 }
+
+#[cfg(test)]
+#[path = "mcp_types_tests.rs"]
+mod tests;
